+++ source/packages/@aws-accelerator/accelerator/lib/stacks/pipeline-stack.ts
@@ -127,36 +127,13 @@ export class PipelineStack extends cdk.Stack {
       maxSessionDuration: cdk.Duration.hours(4),
     });
 
-    // List of regions with AWS CodeStar being supported. For details, see documentation:
-    // https://aws.amazon.com/about-aws/global-infrastructure/regional-product-services/
-    const awsCodeStarSupportedRegions = [
-      'us-east-1',
-      'us-east-2',
-      'us-west-1',
-      'us-west-2',
-      'ap-northeast-2',
-      'ap-southeast-1',
-      'ap-southeast-2',
-      'ap-northeast-1',
-      'ca-central-1',
-      'eu-central-1',
-      'eu-west-1',
-      'eu-west-2',
-      'eu-north-1',
-    ];
-
     new pipeline.AcceleratorPipeline(this, 'Pipeline', {
       toolkitRole,
-      awsCodeStarSupportedRegions,
       ...props,
     });
 
     // cdk-nag suppressions
-    const iam4SuppressionPaths = [
-      'AdminCdkToolkitRole/Resource',
-      'Pipeline/AWSServiceRoleForCodeStarNotifications/ServiceLinkedRoleCodestarNotificationsAmazonawsCom/ServiceLinkedRoleCodestarNotificationsAmazonawsComFunction/ServiceRole/Resource',
-      'Pipeline/AWSServiceRoleForCodeStarNotifications/ServiceLinkedRoleCodestarNotificationsAmazonawsCom/Resource/framework-onEvent/ServiceRole/Resource',
-    ];
+    const iam4SuppressionPaths = ['AdminCdkToolkitRole/Resource'];
 
     const iam5SuppressionPaths = [
       'Pipeline/PipelineRole/DefaultPolicy/Resource',
@@ -190,32 +167,6 @@ export class PipelineStack extends cdk.Stack {
           reason: 'Project requires access to the Docker daemon.',
         },
       ]);
-    }
-
-    // Add NagSuppressions for CodeStar notification in applicable regions
-    if (awsCodeStarSupportedRegions.includes(cdk.Stack.of(this).region)) {
-      // AwsSolutions-IAM4: The IAM user, role, or group uses AWS managed policies.
-      NagSuppressions.addResourceSuppressionsByPath(
-        this,
-        `${this.stackName}/Pipeline/AWSServiceRoleForCodeStarNotifications/CreateServiceLinkedRoleFunction/ServiceRole/Resource`,
-        [
-          {
-            id: 'AwsSolutions-IAM4',
-            reason: 'CodeStar Notification SLR needs managed policies.',
-          },
-        ],
-      );
-      // AwsSolutions-IAM4: The IAM user, role, or group uses AWS managed policies.
-      NagSuppressions.addResourceSuppressionsByPath(
-        this,
-        `${this.stackName}/Pipeline/AWSServiceRoleForCodeStarNotifications/CreateServiceLinkedRoleProvider/framework-onEvent/ServiceRole/Resource`,
-        [
-          {
-            id: 'AwsSolutions-IAM4',
-            reason: 'CodeStar Notification SLR needs managed policies.',
-          },
-        ],
-      );
     }
   }
 }